pyproject.toml
[tool.poetry]
name = "yahoo-agent"
version = "0.1.0"
description = ""
authors = ["Nick Roshdieh <nickroshd@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
langchain = "^0.0.279"
openai = "^0.28.0"
yfinance = "^0.2.28"
pydantic = "^2.3.0"
pandas = "^2.1.0"
python-dotenv = "^1.0.0"
plotly = "^5.16.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
